+++ libs/zhi-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-common
 
+## 1.24.0
+
+### Minor Changes
+
+- fix: change yaml regex
+
+## 1.23.7
+
+### Patch Changes
+
+- feat: Optimize YAML parsing
+
 ## 1.23.6
 
 ### Patch Changes