+++ libs/zhi-github-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # zhi-github-middleware
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-common@1.24.0
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-common@1.23.7
+
 ## 0.4.1
 
 ### Patch Changes