+++ libs/zhi-gitlab-middleware/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-gitlab-middleware
 
+## 0.6.5
+
+### Patch Changes
+
+- zhi-fetch-middleware@0.6.5
+
+## 0.6.4
+
+### Patch Changes
+
+- zhi-fetch-middleware@0.6.4
+
 ## 0.6.3
 
 ### Patch Changes