+++ libs/zhi-xmlrpc-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # zhi-xmlrpc-middleware
 
+## 0.5.12
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-common@1.24.0
+
+## 0.5.11
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-common@1.23.7
+
 ## 0.5.10
 
 ### Patch Changes